refactor: generic hook useSelect
