+++ src/constants/index.ts
@@ -1,0 +1,6 @@
+export const PROGRAMS = [
+    'youtube-dl',
+    'aria2c',
+    'wget'
+] as const
+
