+++ src/hooks/generic/useArray.ts
@@ -1,5 +1,6 @@
 import { useState, useCallback } from 'react';
 
+
 function useArray<T>(initialValue: T[]) {
   const [value, setValue] = useState(initialValue);
